requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
bcrypt==4.2.0
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2